+++ src/config/chatify.php
@@ -38,12 +38,13 @@ return [
     |-------------------------------------
     */
     'pusher' => [
+        'debug' => env('APP_DEBUG', false),
         'key' => env('PUSHER_APP_KEY'),
         'secret' => env('PUSHER_APP_SECRET'),
         'app_id' => env('PUSHER_APP_ID'),
         'options' => [
             'cluster' => env('PUSHER_APP_CLUSTER'),
-            'encrypted' => false,
+            'encrypted' => true,
         ],
     ],
 
@@ -66,7 +67,7 @@ return [
     |-------------------------------------
     */
     'gravatar' => [
-        'enabled' => false,
+        'enabled' => true,
         'image_size' => 200,
         'imageset' => 'identicon'
     ],