+++ CHANGELOG.md
@@ -2,18 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.6.0 (2023-02-xx)
+## v1.6.0 (2023-03-03)
 
 ### Added
 
 - Emoji's support.
 - Css variables.
 - Notification sounds.
+- Auto-time updates.
 
-### Enhancements
+### Changed
 
 - Using UUIDs instead of random IDs on table primary column #243.
 - UI/UX changes and enhancements.
+- Code refactored (part of it).
+- Messenger primary color fallback.
 
 ### Fixed
 
@@ -27,6 +30,7 @@ All notable changes to this project will be documented in this file.
 - Contact list realtime updates issues.
 - Delete messages issues.
 - Fix contact list error `Malformed UTF-8 characters, possibly incorrectly encoded`
+- Search multiple request on typing, debouncing used.
 
 ## v1.5.6 (2023-01-26)